Added more control functionality using new package
Now functions exactly as it did before with old control manager with the added feature of spawning a player on start press. This currently only works with the builder class but will need to be ported over to the healer class as well. Mouse still is not working yet for aiming as well