Map and Gametype select
The arena menu now allows for gametype and map selections. The list will loop back when decrement/increase below/above the allowed values.
The text for the maps default with the first in the relevant enums.
Moving forwards, as more gametypes and maps are added, add the names to the enum.